# docs/swagger.yaml
basePath: /
definitions:
  app.AddonParams:
    properties:
      cancelURL:
        type: string
      email:
        type: string
      priceID:
        type: string
      quantity:
        type: integer
      successURL:
        type: string
      userID:
        type: string
    type: object
  app.Address:
    properties:
      city:
        type: string
      country:
        type: string
      line1:
        type: string
      postalCode:
        type: string
      state:
        type: string
    type: object
  app.CheckoutParams:
    properties:
      address:
        $ref: '#/definitions/app.Address'
      cancelURL:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      planID:
        type: string
      stripeSubscriptionID:
        type: string
      successURL:
        type: string
      timezone:
        type: string
      userID:
        type: string
    type: object
  command.SetupFreeTierParams:
    properties:
      address:
        $ref: '#/definitions/app.Address'
      country:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      timezone:
        type: string
      userId:
        type: string
    type: object
host: localhost:9003
info:
  contact: {}
  description: API for managing subscriptions, AI credits, and Stripe integration.
  title: Teaching Assistant Billing Service API
  version: "2.0"
paths:
  /api/v2/billing/addon:
    post:
      consumes:
      - application/json
      description: Creates a new Stripe checkout session for a storage add-on
      parameters:
      - description: Add-on parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.AddonParams'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the checkout URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a Stripe session for an add-on
      tags:
      - billing
  /api/v2/billing/balance:
    get:
      description: Returns the current AI credit balance for a user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns the wallet balance
          schema:
            additionalProperties: true
            type: object
        "400":
          description: userId is required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user wallet balance
      tags:
      - billing
  /api/v2/billing/check-usage:
    get:
      description: Verifies if a user has enough credits/storage and deducts if necessary
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: Usage type (ai_tokens, storage)
        in: query
        name: type
        required: true
        type: string
      - description: 'Amount to check/deduct (default: 1)'
        in: query
        name: amount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Status OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "402":
          description: Usage limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check and deduct usage
      tags:
      - billing
  /api/v2/billing/checkout:
    post:
      consumes:
      - application/json
      description: Creates a new Stripe checkout session for a subscription plan or
        AI credit pack
      parameters:
      - description: Checkout parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.CheckoutParams'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the checkout URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a Stripe checkout session
      tags:
      - billing
  /api/v2/billing/portal:
    post:
      description: Creates a session for the Stripe customer portal where users can
        manage subscriptions
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: URL to return to after leaving the portal
        in: query
        name: returnUrl
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns the portal URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: userId is required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a Stripe customer portal session
      tags:
      - billing
  /api/v2/billing/setup-free-tier:
    post:
      consumes:
      - application/json
      description: Automatically assigns a Free tier subscription to a new user
      parameters:
      - description: Setup parameters (email, name, address, etc)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/command.SetupFreeTierParams'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Setup free tier for a new user
      tags:
      - billing
  /api/v2/billing/sync-customer:
    post:
      consumes:
      - application/json
      description: Synchronizes user profile changes (name, email, address, etc) with
        Stripe
      parameters:
      - description: Sync parameters
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Customer not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync customer data with Stripe
      tags:
      - billing
swagger: "2.0"